+++ grails-app/controllers/io/xh/hoist/admin/ConfigAdminController.groovy
@@ -8,9 +8,9 @@
 package io.xh.hoist.admin
 
 import io.xh.hoist.config.AppConfig
-import io.xh.hoist.security.Access
+import io.xh.hoist.security.AccessRequiresRole
 
-@Access(['HOIST_ADMIN_READER'])
+@AccessRequiresRole('HOIST_ADMIN_READER')
 class ConfigAdminController extends AdminRestController {
 
     static restTarget = AppConfig