+++ src/main/groovy/io/xh/hoist/exception/NotAuthorizedException.groovy
@@ -13,7 +13,7 @@ import static org.apache.hc.core5.http.HttpStatus.SC_FORBIDDEN
  * Exception for use when the authenticated user does not have access to the resource in question.
  *
  * This exception is thrown by Hoist's by {@link io.xh.hoist.security.AccessInterceptor} if an
- * authenticated user does not have roles required by a controller's `@Access` annotation.
+ * authenticated user does not have roles required by a controller's `@AccessXXX` annotation.
  *
  * Applications may also throw this exception, or subclasses of it, directly in response to requests
  * they cannot fulfill due to authorization-related constraints.