+++ src/main/groovy/io/xh/hoist/security/AccessRequiresAllRoles.groovy
@@ -2,7 +2,7 @@
  * This file belongs to Hoist, an application development toolkit
  * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
  *
- * Copyright © 2026 Extremely Heavy Industries Inc.
+ * Copyright © 2025 Extremely Heavy Industries Inc.
  */
 
 package io.xh.hoist.security
@@ -10,11 +10,13 @@ package io.xh.hoist.security
 import java.lang.annotation.*
 
 /**
- * Controller annotation to indicate that any user can access, regardless of roles.
+ * Controller annotation to list roles required to execute any action.
+ * Current user must have ALL roles listed to access.
  */
 @Inherited
 @Target([ElementType.METHOD, ElementType.TYPE])
 @Retention(RetentionPolicy.RUNTIME)
-@interface AccessAll {
-
+@interface AccessRequiresAllRoles {
+    /** Array of role names, all of which are required for access. */
+    String[] value()
 }